backend/config.py
import os
from dotenv import load_dotenv

load_dotenv()  # Load environment variables from .env file

GOOGLE_API_KEY = os.getenv("GOOGLE_API_KEY")
MODEL = os.getenv("GOOGLE_MODEL")
FALLBACK_MODEL = os.getenv("FALLBACK_MODEL")
