backend/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
python-pptx==1.0.2
python-dotenv==1.1.1
pydantic==2.11.7
openai==2.30.0
google-genai==1.70.0
manim==0.20.1
